+++ raiden-dapp/src/components/tokens/TokenList.vue
@@ -1,0 +1,50 @@
+<template>
+  <div class="token-list d-flex flex-column">
+    <list-header v-if="header" :header="header" class="flex-grow-0" />
+
+    <div class="token-list__wrapper flex-grow-1">
+      <v-virtual-scroll
+        #default="{ item }"
+        :items="tokens"
+        item-height="105"
+        bench="400"
+      >
+        <token-list-item :token="item" @select="forwardTokenSelection" />
+      </v-virtual-scroll>
+    </div>
+  </div>
+</template>
+
+<script lang="ts">
+import { Component, Vue, Prop, Emit } from 'vue-property-decorator';
+import { Token } from '@/model/types';
+import ListHeader from '@/components/ListHeader.vue';
+import TokenListItem from './TokenListItem.vue';
+
+@Component({ components: { ListHeader, TokenListItem } })
+export default class TokenList extends Vue {
+  @Prop({ type: String })
+  header!: string;
+
+  @Prop({ required: true })
+  tokens!: Token[];
+
+  @Emit('select')
+  forwardTokenSelection(token: Token): Token {
+    return token;
+  }
+}
+</script>
+
+<style lang="scss" scoped>
+@import '@/scss/scroll';
+
+.token-list {
+  height: 100%;
+
+  &__wrapper {
+    overflow-y: auto;
+    @extend .themed-scrollbar;
+  }
+}
+</style>